aws_resources/sagemaker/requirements.txt
# YOLOv11 and dependencies
# YOLOv11 requires ultralytics >= 8.3.0 (C3k2 module support)
ultralytics>=8.3.0

# PyTorch (CUDA 11.8 compatible)
torch==2.1.0
torchvision==0.16.0

# Image processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# Web server
flask==3.0.0
gunicorn==21.2.0

# Utilities
numpy==1.24.3
